Simplified Operators
This simplifies the Operator defintion system for ModularAgents and agnets in general. feature_set and function_set must now be provided as part of the args objec for a ModularAgent rather than at the top level. Further, Operators for the FO_planner must be pre-registered and refered to by name in the feature_set and function_set parameters. All other approaches to defining operators will no longer work and will provide warnings in the HTTP response if they are used.