Increased Warnings
Added a bunch of warning and error statements to views for things that are not totally working right now. I opted to warn about saving and loading Agents rather than remove the serialization system entirely as we will want it to work mostly as written some time in the future.