Made several improvements to the tast_save.
Removed redundant tests and tests that were only evaluating behavior of external libraries not related to our code.
Fixed up the fo_planner test so that is equivalent to the numbert one, with the caveat that it does not verify structural equivalence because of lambdas.
Removed dead code and cleaned up code style to bring it closer to PEP 8 spec.