Changed requirements file to include torch (#80)
* Changed requirements file to include torch

It appears that torch now hosts a standardly named version of their windows library on pip, making it possible to include torch in the requirements file without an extra step. This makes that change and updates the README accordingly.

* Removed torch install from the travis.yml

Forgot to remove it from the traivs.yml since its in the core requirements now. Shouldn't affect anything but want to make sure the test is true.